fix(review): H-ARCH-2 Document.amount auf Decimal umgestellt
- Document.amount: Mapped[Decimal | None] statt float. SQLAlchemy mapped
  Numeric(12,2) jetzt auf Decimal — kein impliziter IEEE-754-Roundtrip.
- Pydantic-Schemas (DocumentResponse, DocumentListItem, DocumentUpdate,
  SearchResultItem, TaxCategorySummary, TaxYearSummary, TaxDocumentItem):
  amount/total_amount auf Decimal. Pydantic v2 serialisiert das standard-
  maessig als JSON-String — Frontend formatAmount() macht parseFloat, also
  ohne Anpassung kompatibel.
- tax_export_service.get_tax_summary: Aggregation mit Decimal statt float
  + Quantisierung auf 2 Nachkommastellen via ROUND_HALF_UP. Summen ueber
  1000 Belege sind jetzt mathematisch korrekt.

PDF/CSV-Generierung (Anzeige) bleibt float-basiert — fuer einzelne
Display-Werte ist der Rundungsfehler vernachlaessigbar; die persistierten
Aggregate sind jetzt sauber.

358 Backend-Tests passed, 1 skipped.

+++ backend/app/models/document.py
@@ -1,6 +1,7 @@
 import enum
 import uuid
 from datetime import date, datetime, timezone
+from decimal import Decimal
 
 from sqlalchemy import (
     Boolean,
@@ -84,7 +85,10 @@ class Document(Base):
     )
     title: Mapped[str] = mapped_column(String(500), nullable=False, default="Unbekanntes Dokument")
     document_date: Mapped[date | None] = mapped_column(Date, nullable=True, index=True)
-    amount: Mapped[float | None] = mapped_column(Numeric(12, 2), nullable=True)
+    # H-ARCH-2: Decimal statt float — vermeidet IEEE-754-Rundungsfehler bei
+    # Aggregationen (Tax-Year-Summary, Sonderausgaben). DB-Storage ist
+    # Numeric(12,2), SQLAlchemy mapped es jetzt nach Decimal statt float.
+    amount: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
     currency: Mapped[str] = mapped_column(String(3), nullable=False, default="EUR")
     issuer: Mapped[str | None] = mapped_column(String(500), nullable=True, index=True)
     recipient: Mapped[str | None] = mapped_column(String(500), nullable=True)
